half way through changing up api
